Move leaderboard fragments into their own package
