+++ app/src/main/java/com/qrcode_quest/ui/leaderboard/PlayerListFragment.java
@@ -1,24 +1,21 @@
-package com.qrcode_quest;
+package com.qrcode_quest.ui.leaderboard;
 
 
 import static com.qrcode_quest.ui.leaderboard.LeaderboardFragmentDirections.*;
 
 import android.os.Bundle;
 
-import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 import androidx.navigation.NavController;
-import androidx.navigation.Navigation;
 import androidx.navigation.fragment.NavHostFragment;
 
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
 import android.widget.ListView;
 
+import com.qrcode_quest.R;
 import com.qrcode_quest.entities.PlayerAccount;
 import com.qrcode_quest.ui.leaderboard.LeaderboardFragmentDirections;
 import com.qrcode_quest.ui.leaderboard.LeaderboardFragmentDirections.ActionNavigationLeaderboardToNavigationPlayerQrlist;
